requirements.txt
# Core Discord Bot Dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0

# AI/ML Dependencies  
torch>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# ComfyUI Integration
websockets>=11.0.0

# Neo4j Database
neo4j>=5.12.0

# Optional: For enhanced Neo4j functionality
# py2neo>=2021.2.3